requirements.txt
mindsdb-sdk>=3.4.4
typer[all]>=0.16.0
pypdf>=5.7.0
rich>=14.0.0
tqdm>=4.67.1
google-generativeai>=0.8.5
flask>=3.1.1
pyyaml>=6.0.2
requests>=2.31.0
beautifulsoup4>=4.12.0
ollama>=0.3.0
